Add code to properly orient glyphs of surf mols and vol mols in Blender > 3.3
Blender version 3.3 and above break the user setting of vertex normals.
This code is a workaround for that breakage.  Here we make an
oriented plane for each molecule and use the face normal to orient
the instanced glyphs.

Unfortunately the performance is 20x slower!

In the next commit we will add options to turn off orientation calculations
for surf mols and vol mols to regain good performance when needed.